requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cloudinary==1.24.0
cryptography==3.3.1
defusedxml==0.7.0rc2
distlib==0.3.1
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.1.0
pipenv==2020.11.15
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.3.1
virtualenv-clone==0.5.4
